requirements.txt
Flask
Pillow
colorgram.py
beautifulsoup4
lxml
requests
Jinja2
cssutils
python-dotenv